[File]: Providers for Each Widgets
- 위젯화면별 Provider를 생성
- 내부코드는 동일하며 추후 수정예정
